+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Fix job alert emails
+
 ## 17.0 (2020-05-21)
 
 Compare: https://git.torchbox.com/buckinghamshire-council/bc/compare/16.1...17.0